pttk-scraper/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
joblib==1.0.0
nltk==3.5
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
soupsieve==2.1
tqdm==4.55.1
urllib3==1.26.2
